# app/backend/api-spec.yaml
openapi: 3.1.0
info:
  title: user-manager
  description: ''
  contact:
    name: FLECS Technologies GmbH
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.2.1
paths:
  /meta/issuer:
    get:
      tags:
      - Experimental
      operationId: get
      responses:
        '200':
          description: Issuer that issues auth tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uri'
  /meta/jwk:
    get:
      tags:
      - Experimental
      operationId: get
      responses:
        '200':
          description: Jwk that has to be used to verify issued tokens
          content:
            application/json:
              schema: {}
  /users:
    get:
      tags:
      - rest::users
      operationId: get
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
    put:
      tags:
      - rest::users
      operationId: put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '201':
          description: User was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/u16'
        '400':
          description: Invalid request body
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: User with that name already exists
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /users/self:
    delete:
      tags:
      - rest::users::self_
      operationId: delete
      responses:
        '204':
          description: User was deleted
        '401':
          description: Not authenticated
        '403':
          description: Cannot delete the super admin
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /users/super-admin:
    get:
      tags:
      - Experimental
      operationId: get
      responses:
        '204':
          description: Super admin exists
        '404':
          description: Super admin does not exist
    post:
      tags:
      - Experimental
      operationId: post
      requestBody:
        description: The super admin that should be set
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuperAdmin'
        required: true
      responses:
        '200':
          description: Super admin was created
        '400':
          description: Invalid password
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: Super admin already exists
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /users/{uid}:
    delete:
      tags:
      - rest::users::uid
      operationId: delete
      parameters:
      - name: uid
        in: path
        description: User ID to delete
        required: true
        schema:
          $ref: '#/components/schemas/u16'
      responses:
        '204':
          description: User was deleted
        '400':
          description: Invalid user ID
        '403':
          description: Cannot delete the super admin
        '404':
          description: User does not exist
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    CreateUser:
      type: object
      required:
      - name
      - password
      - groups
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupId'
          uniqueItems: true
        name:
          type: string
        password:
          type: string
    GroupId:
      type: string
    SuperAdmin:
      type: object
      required:
      - name
      - full_name
      - password
      properties:
        full_name:
          type: string
        name:
          type: string
        password:
          type: string
    Uri:
      type: string
      format: uri
    UserSummary:
      type: object
      required:
      - id
      - name
      - groups
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupId'
          uniqueItems: true
        id:
          $ref: '#/components/schemas/u16'
        name:
          type: string
    u16:
      type: integer
      format: int32
      minimum: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []