# app/backend/src/rest/meta/issuer.rs
use crate::state;
use axum::Json;
use axum::extract::State;
use axum::http::StatusCode;
use axum::response::{IntoResponse, Response};
use serde::Serialize;
use utoipa::openapi::{RefOr, Schema};
use utoipa::{PartialSchema, ToSchema};

#[derive(Serialize)]
struct Uri(url::Url);

impl PartialSchema for Uri {
    fn schema() -> RefOr<Schema> {
        utoipa::openapi::schema::ObjectBuilder::new()
            .schema_type(utoipa::openapi::schema::Type::String)
            .format(Some(utoipa::openapi::SchemaFormat::KnownFormat(
                utoipa::openapi::KnownFormat::Uri,
            )))
            .into()
    }
}

impl ToSchema for Uri {}

#[utoipa::path(
    get,
    path="/meta/issuer",
    tag = "Experimental",
    responses(
        (status = OK, description = "Issuer that issues auth tokens", body = Uri)
    )
)]
pub async fn get(State(state): State<state::AppState>) -> Response {
    let issuer = state.issuer.lock().unwrap().url.clone();
    (StatusCode::OK, Json(Uri(issuer))).into_response()
}